Add user services and stuff

+++ app/controllers/users_controller.rb
@@ -106,15 +106,17 @@ class UsersController < ApplicationController
 	
 	def parse_user_types
 		#I'm ashamed of this code. I really am. But can't do it with strong params yet. I'll work on that.
-		params[:user][:user_types_attributes].each do |key, value|
-			if value[:user_type] == "0"
-				UserType.where(user_id: @user.id, user_type: UserType.available_types[key.to_i][0]).destroy_all
-			else
-				if value[:user_type] == UserType.available_types[key.to_i][0] && UserType.where(user_id: @user.id, user_type: value[:user_type]).empty?
-					@user.user_types.create(user_type: value[:user_type])
+		if params[:user][:user_types_attributes]
+			params[:user][:user_types_attributes].each do |key, value|
+				if value[:user_type] == "0"
+					UserType.where(user_id: @user.id, user_type: UserType.available_types[key.to_i][0]).destroy_all
+				else
+					if value[:user_type] == UserType.available_types[key.to_i][0] && UserType.where(user_id: @user.id, user_type: value[:user_type]).empty?
+						@user.user_types.create(user_type: value[:user_type])
+					end
 				end
 			end
+			UserType.where(user_type: "0").destroy_all
 		end
-		UserType.where(user_type: "0").destroy_all
 	end
 end
